requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
fuzzywuzzy==0.18.0
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-Levenshtein==0.12.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
